requirements.txt
selenium~=4.21.0
pytest~=8.2.0
Faker~=26.0.0
allure-pytest==2.13.5